chore(gameCardOptions): move to app folder
